chore: create utility getAwsAuthSigner
